# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kumquat API
  version: 1.0.0
  description: Northbound business API. Engine custom resources are internal execution projections.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/Size'}]
      responses: {'200': {$ref: '#/components/responses/UserList'}}
    post:
      operationId: createUser
      requestBody: {$ref: '#/components/requestBodies/UserCreate'}
      responses: {'200': {$ref: '#/components/responses/User'}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /users/{id}:
    parameters: [{$ref: '#/components/parameters/NumericID'}]
    get: {operationId: getUser, responses: {'200': {$ref: '#/components/responses/User'}, '404': {$ref: '#/components/responses/Error'}}}
    put:
      operationId: updateUser
      requestBody: {$ref: '#/components/requestBodies/UserUpdate'}
      responses: {'200': {$ref: '#/components/responses/User'}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
    delete: {operationId: deleteUser, responses: {'200': {$ref: '#/components/responses/Message'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    NumericID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
...
    Page: {name: page, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}}
    Size: {name: size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  requestBodies:
...
    UserCreate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserCreate'}}}}
    UserUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserUpdate'}}}}
...
  schemas:
    PublicID: {type: string, minLength: 3, maxLength: 40, pattern: '^[a-z]+_[a-f0-9]+$'}
...
    UserCreate: {type: object, additionalProperties: false, required: [username, email, password, role_id], properties: {username: {type: string, minLength: 3, maxLength: 32}, email: {type: string, format: email}, password: {type: string, minLength: 6, maxLength: 32}, nickname: {type: string}, role_id: {type: integer, minimum: 1}, module_id: {type: integer, minimum: 1, nullable: true}}}
    UserUpdate: {type: object, additionalProperties: false, properties: {nickname: {type: string}, role_id: {type: integer, minimum: 1}, module_id: {type: integer, minimum: 1, nullable: true}}}
...
    ErrorEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer}, message: {type: string}}}
    RoleDTO: {type: object, additionalProperties: false, required: [id, name, created_at, updated_at], properties: {id: {type: integer, format: int64}, name: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    ModuleDTO: {type: object, additionalProperties: false, required: [id, name, level, sort, path, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/PublicID'}, name: {type: string}, parent_id: {$ref: '#/components/schemas/PublicID'}, level: {type: integer}, sort: {type: integer}, path: {type: string}, children: {type: array, items: {$ref: '#/components/schemas/ModuleDTO'}}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    UserDTO: {type: object, additionalProperties: false, required: [id, username, email, nickname, role_id, created_at, updated_at], properties: {id: {type: integer, format: int64}, username: {type: string}, email: {type: string, format: email}, nickname: {type: string}, role_id: {type: integer, format: int64}, role: {$ref: '#/components/schemas/RoleDTO'}, module_id: {type: integer, format: int64, nullable: true}, module: {$ref: '#/components/schemas/ModuleDTO'}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
...
    Pagination: {type: object, additionalProperties: false, required: [total, page, size, list], properties: {total: {type: integer, format: int64}, page: {type: integer}, size: {type: integer}, list: {type: array, items: {type: object}}}}
    MessageEnvelope: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: integer, enum: [0]}, message: {type: string}}}
...
    UserEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {$ref: '#/components/schemas/UserDTO'}}}
...
    UserListEnvelope: {type: object, additionalProperties: false, required: [code, message, data], properties: {code: {type: integer, enum: [0]}, message: {type: string}, data: {allOf: [{$ref: '#/components/schemas/Pagination'}, {type: object, properties: {list: {type: array, items: {$ref: '#/components/schemas/UserDTO'}}}}]}}}
...
  responses:
...
    User: {description: User, content: {application/json: {schema: {$ref: '#/components/schemas/UserEnvelope'}}}}
    UserList: {description: Users, content: {application/json: {schema: {$ref: '#/components/schemas/UserListEnvelope'}}}}
...
    Message: {description: Message, content: {application/json: {schema: {$ref: '#/components/schemas/MessageEnvelope'}}}}
...
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}